+++ spec/support/entities/admin_user.rb
@@ -1,0 +1,5 @@
+require_relative 'user'
+
+class AdminUser < User
+  attribute :level, Integer
+end