+++ spec/support/repositories/active_record/admin_user_repository.rb
@@ -1,0 +1,5 @@
+require_relative 'user_repository'
+
+class AdminUserRepositoryActiveRecord < UserRepositoryActiveRecord
+  for_entity AdminUser
+end