+++ spec/support/repositories/sequel/admin_user_repository.rb
@@ -1,0 +1,5 @@
+require_relative 'user_repository'
+
+class AdminUserRepositorySequel < UserRepositorySequel
+  for_entity AdminUser
+end